Refactor : Modal 컴포넌트 폴더 구조 변경 - modal 기능 관련 코드는 하나의 파일로 묶고, Context 폴더로 옮김.
