+++ src/app/(content-layout)/[groupId]/editgroup/_editgroup/DeleteGroupButton.tsx
@@ -5,14 +5,14 @@ import { useState } from 'react';
 import { Toast } from '@/components/common/Toastify';
 import DangerModal from '@/components/danger-modal';
 import BouncingDots from '@/components/common/loading/BouncingDots';
-import useModalContext from '@/components/common/modal/core/useModalContext';
+import { useModal } from '@/contexts/ModalContext';
 import TrashCan from '@/assets/TrashCan';
 import { deleteGroup } from '../action';
 import { useUser } from '@/contexts/UserContext';
 
 export default function DeleteGroupButton({ groupId }: { groupId: number }) {
   const [isPending, setIsPending] = useState(false);
-  const { openModal } = useModalContext();
+  const { openModal } = useModal();
   const { fetchUser } = useUser();
   const router = useRouter();
