+++ src/app/(content-layout)/[groupId]/tasklist/_tasklist/components/manage-task-item-modal/MangeTaskItemModal.tsx
@@ -1,4 +1,5 @@
-import { ModalContainer, ModalOverlay, ModalPortal, ModalTrigger } from '@/components/common/modal';
+import { ModalContainer, ModalOverlay, ModalTrigger } from '@/components/common/modal';
+import { ModalPortal } from '@/contexts/ModalContext';
 import ManageTaskItem from '@/components/manage-task-item/components/ManageTaskItem';
 import Plus from '@/assets/Plus';
 import { TaskItemProps } from '@/components/manage-task-item/type';
