+++ src/app/mypage/_mypage/mypage-modal/ConfirmDeleteAccountModal.tsx
@@ -8,16 +8,15 @@ import {
   ModalFooter,
   ModalHeading,
   ModalOverlay,
-  ModalPortal,
 } from '@/components/common/modal';
-import useModalContext from '@/components/common/modal/core/useModalContext';
+import { useModal, ModalPortal } from '@/contexts/ModalContext';
 import axiosClient from '@/lib/axiosClient';
 import Button from '@/components/common/Button';
 import { Toast } from '@/components/common/Toastify';
 import { deleteClientCookie } from '@/lib/cookie/client';
 
 export default function ConfirmDeleteAccountModal() {
-  const { closeModal } = useModalContext();
+  const { closeModal } = useModal();
   const { logoutUser } = useUser();
 
   return (
