+++ src/components/signup-alert-modal/SignupFailModal.tsx
@@ -7,13 +7,12 @@ import {
   ModalFooter,
   ModalHeading,
   ModalOverlay,
-  ModalPortal,
 } from '@/components/common/modal';
+import { useModal, ModalPortal } from '@/contexts/ModalContext';
 import Button from '../common/Button';
-import useModalContext from '@/components/common/modal/core/useModalContext';
 
 export default function SignupFailModal() {
-  const { closeModal } = useModalContext();
+  const { closeModal } = useModal();
 
   return (
     <ModalPortal modalId="signup-fail">
